Development (#64)
* fix warnings

* IconFonts and ImGui version bump

* Gui works

* Warning fix

* Switch to FontAwesome6

System names created directly

* move starsystem name/pos generation to component

* Gui work

* Options menu with galaxyConfig binding

* cached Markov name map

pre-generated/cached star names

* re-added marker rendering

fine-grained RenderSystem config

* Options menu for number of stars

using GalaxyConfig to transport options from menu to game stage

* gui-refactoring

* gui-refactoring

* Warning fixes

* attempted warning setup for 3rd party

* refactoring menu system

* renaming and include hygiene

* renaming and include hygiene

* Fix Asteroids game

* more cleaning

* move options menu into separate file

* options overhaul

* config improvements

tabbed config
change detection

* ScopeBeginEndWrapper utility

* renamed guiElements

* move style stack and ImGui BeginEnd wrappers to separate file

* using ScopedBeginEnd wrappers

* minimal typeRegistrar documentation

* Fixed ScopeBeginEndWrapper ParameterInstance to allow generic types

Added BeginChild wrapper

* Void return variant for ScopeBeginEndWrapper

* OptionsMenu with factions

* faction config and handling

* cleanup and fixed menu background music

---------

+++ libs/pgEngine/math/Random.hpp
@@ -15,7 +15,7 @@ public:
     {
     }
 
-    SeedGenerator(uint64_t fixed_seed)
+    SeedGenerator(uint32_t fixed_seed)
       : _gen(fixed_seed)
     {
     }
